requirements.txt
flask==2.3.3
openai==0.27.8
python-dotenv==1.0.1
flask-cors==4.0.0
markdown==3.5.1
mdx-truly-sane-lists==1.3
pymdown-extensions==10.7 